[WFLY-410] Add in transformers for legacy models.
