+++ clustering/jgroups/src/main/java/org/jboss/as/clustering/jgroups/subsystem/JGroupsExtension.java
@@ -21,9 +21,6 @@
  */
 package org.jboss.as.clustering.jgroups.subsystem;
 
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.ADD;
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.WRITE_ATTRIBUTE_OPERATION;
-
 import java.security.AccessController;
 import java.security.PrivilegedAction;
 import java.util.List;
@@ -31,7 +28,6 @@ import java.util.List;
 import org.jboss.as.clustering.jgroups.LogFactory;
 import org.jboss.as.controller.Extension;
 import org.jboss.as.controller.ExtensionContext;
-import org.jboss.as.controller.ModelVersion;
 import org.jboss.as.controller.PathElement;
 import org.jboss.as.controller.SubsystemRegistration;
 import org.jboss.as.controller.descriptions.ModelDescriptionConstants;
@@ -39,9 +35,6 @@ import org.jboss.as.controller.descriptions.ResourceDescriptionResolver;
 import org.jboss.as.controller.descriptions.StandardResourceDescriptionResolver;
 import org.jboss.as.controller.parsing.ExtensionParsingContext;
 import org.jboss.as.controller.registry.ManagementResourceRegistration;
-import org.jboss.as.controller.transform.RejectExpressionValuesTransformer;
-import org.jboss.as.controller.transform.ResourceTransformer;
-import org.jboss.as.controller.transform.TransformersSubRegistration;
 import org.jboss.dmr.ModelNode;
 import org.jboss.staxmapper.XMLElementReader;
 import org.jgroups.Global;
@@ -105,7 +98,7 @@ public class JGroupsExtension implements Extension {
 
         if (context.isRegisterTransformers()) {
             // Register the model transformers
-            registerTransformers(registration);
+            JGroupsTransformers.registerTransformers(registration);
         }
     }
 
@@ -121,67 +114,5 @@ public class JGroupsExtension implements Extension {
                 context.setSubsystemXmlMapping(SUBSYSTEM_NAME, namespace.getUri(), reader);
             }
         }
-    }
-
-    // Transformation
-
-    /**
-     * Register the transformers for older model versions.
-     *
-     * @param subsystem the subsystems registration
-     */
-    private static void registerTransformers(final SubsystemRegistration subsystem) {
-        registerTransformers_1_1_0(subsystem);
-        registerTransformers_1_2_0(subsystem);
-    }
-
-    private static void registerTransformers_1_2_0(final SubsystemRegistration subsystem) {
-        final ModelVersion version = ModelVersion.create(1, 2, 0);
-
-        final TransformersSubRegistration registration = subsystem.registerModelTransformers(version, ResourceTransformer.DEFAULT);
-        final TransformersSubRegistration stack = registration.registerSubResource(StackResourceDefinition.STACK_PATH);
-
-        registerRelayTransformers(stack);
-    }
-
-    private static void registerTransformers_1_1_0(final SubsystemRegistration subsystem) {
-        // Transformations to the 1.1.0 Model
-        // - we need to reject expressions for transport (and similarly for protocol properties) for these operations
-        //   transport=TRANSPORT/property=<name>:add(value=<value>)
-        //   transport=TRANSPORT/property=<name>:write-attribute(name=value, value=<value>)
-        //   transport=TRANSPORT:add(...,properties=<list of properties>)
-
-        final ModelVersion version = ModelVersion.create(1, 1, 0);
-        final RejectExpressionValuesTransformer transformer = new RejectExpressionValuesTransformer(PropertyResourceDefinition.VALUE,
-                TransportResourceDefinition.PROPERTIES, ProtocolResourceDefinition.PROPERTIES, TransportResourceDefinition.SHARED);
-        final ResourceTransformer resourceTransformer = transformer;
-
-        final TransformersSubRegistration registration = subsystem.registerModelTransformers(version, ResourceTransformer.DEFAULT);
-        final TransformersSubRegistration stack = registration.registerSubResource(StackResourceDefinition.STACK_PATH);
-
-        // reject expressions for transport properties, for the add and write-attribute op
-        final TransformersSubRegistration transport = stack.registerSubResource(TransportResourceDefinition.TRANSPORT_PATH, resourceTransformer);
-        transport.registerOperationTransformer(ADD, transformer);
-        transport.registerOperationTransformer(WRITE_ATTRIBUTE_OPERATION, transformer.getWriteAttributeTransformer());
-        final TransformersSubRegistration transport_property = transport.registerSubResource(PropertyResourceDefinition.PROPERTY_PATH) ;
-        transport_property.registerOperationTransformer(ADD, transformer);
-        transport_property.registerOperationTransformer(WRITE_ATTRIBUTE_OPERATION, transformer.getWriteAttributeTransformer());
-
-        // reject expressions for transport properties, for the add and write-attribute op
-        final TransformersSubRegistration protocol = stack.registerSubResource(ProtocolResourceDefinition.PROTOCOL_PATH, resourceTransformer);
-        protocol.registerOperationTransformer(ADD, transformer);
-        final TransformersSubRegistration protocol_property = protocol.registerSubResource(PropertyResourceDefinition.PROPERTY_PATH);
-        protocol_property.registerOperationTransformer(ADD, transformer);
-        protocol_property.registerOperationTransformer(WRITE_ATTRIBUTE_OPERATION, transformer.getWriteAttributeTransformer());
-
-        registerRelayTransformers(stack);
-    }
-
-    private static void registerRelayTransformers(final TransformersSubRegistration stack) {
-        final TransformersSubRegistration relay = stack.registerSubResource(RelayResource.PATH, true);
-        relay.discardOperations(ADD, WRITE_ATTRIBUTE_OPERATION);
-
-        final TransformersSubRegistration remoteSite = relay.registerSubResource(RemoteSiteResource.PATH, true);
-        remoteSite.discardOperations(ADD, WRITE_ATTRIBUTE_OPERATION);
     }
 }
