+++ clustering/jgroups/src/test/resources/org/jboss/as/clustering/jgroups/subsystem/jgroups-transformer_1_1.xml
@@ -19,5 +19,6 @@
     </stack>
     <stack name="minimal">
         <transport type="UDP"/>
+        <protocol type="MPING"/>
     </stack>
 </subsystem>
