+++ clustering/jgroups/src/test/resources/org/jboss/as/clustering/jgroups/subsystem/jgroups-transformer_1_2.xml
@@ -23,5 +23,6 @@
     </stack>
     <stack name="minimal">
         <transport type="UDP"/>
+        <protocol type="MPING"/>
     </stack>
 </subsystem>
